DBZ-6182 Simplifying inheritance hierarchy in MongoDB tests
